# optimize the sum calculation of multiple inputs