@Aurelius84 2020-01-03
PR导致test_optimizer_grad.py单测挂掉，观测的现象如下：
在`TestAdamOptimizer`这个单测函数里，参数`x`的输出值是1.，梯度更新后应该0.99=1. - 0.01，但此PR下返回是1.01 = 1. + 0.01。看起来是sum求和时，符号反了。
```
[2020-01-02 15:53:06] AssertionError: 
[2020-01-02 15:53:06] Not equal to tolerance rtol=1e-07, atol=0
[2020-01-02 15:53:06] 
[2020-01-02 15:53:06] Mismatched elements: 160 / 160 (100%)
[2020-01-02 15:53:06] Max absolute difference: 0.01999903
[2020-01-02 15:53:06] ax relative difference: 0.02020103
[2020-01-02 15:53:06]  x: array([[1.01, 1.01, 1.01, 1.01, 1.01, 1.01, 1.01, 1.01, 1.01, 1.01],
[2020-01-02 15:53:06]        [1.01, 1.01, 1.01, 1.01, 1.01, 1.01, 1.01, 1.01, 1.01, 1.01],
[2020-01-02 15:53:06]        [1.01, 1.01, 1.01, 1.01, 1.01, 1.01, 1.01, 1.01, 1.01, 1.01],...
[2020-01-02 15:53:06]  y: array([[0.99, 0.99, 0.99, 0.99, 0.99, 0.99, 0.99, 0.99, 0.99, 0.99],
[2020-01-02 15:53:06]        [0.99, 0.99, 0.99, 0.99, 0.99, 0.99, 0.99, 0.99, 0.99, 0.99],
[2020-01-02 15:53:06]        [0.99, 0.99, 0.99, 0.99, 0.99, 0.99, 0.99, 0.99, 0.99, 0.99],...
[2020-01-02 15:53:06] 
[2020-01-02 15:53:06] ======================================================================
```

### 复现
可以在test_optimizer_grad.py里的第219行添加log信息：
```
param_grads = gt_params + gt_grads
                                for i in range(len(res)):
                                    print("val of i: ", i)
                                    print(res[i])   # 框架计算值
                                    print(param_grads[i])   # 真实值
                                    np.testing.assert_allclose(res[i],
                                                               param_grads[i])
```